use temp for testing media dir (deals with vagrant shared mount problems)

+++ neurovault/apps/statmaps/tests/test_compressed.py
@@ -40,7 +40,8 @@ class UploadFolderTestCase(TestCase):
  
     def tearDown(self):
         clearDB()
-        shutil.rmtree(self.tmpdir)
+        if os.path.exists(self.tmpdir):
+            shutil.rmtree(self.tmpdir)
         self.coll.delete()
         self.user.delete()
